Implemented CSS styles
The buttons and links are formatted so that they look more appealing and pretty to viewers. This is still a work in progress, so maybe I'll have to play around with the properties until I am satisfied with the CSS styles.
Also, for pages that will show images (still have to implement them), the forward and backward buttons are added as well.