Added images
In the next few weeks, I will try to implement the LightBox gallery so i can easily display the images within the single page.
Also, the image gallery section for CPSC 120 and 585 are deleted, because for the former, the images doesn't look good enough and appealing, and for the latter, it would be troublesome to obtain some screenshots for it.