Implemented slideshow gallery
To ensure that the images are displayed properly, the width and height ratios are set to 100%.
For CPSC 121, only eight images are implemented for this page as there are 24 screenshots included in the images folder, which is too much to handle.
Also, to summarize which .css file is used, webpages with screenshots will use the styles_gallery.css file whereas those without will still use styles_default.css file.