Deleted image gallery
The implementation of Bootstrap is causing issues with the image gallery, meaning that screenshots of sample applications did not display correctly. Let's get rid of the images and the folder containing the screenshots for now. (The JS file for the image gallery will remain at is for now. Perhaps in the future, it will be reimplemented.)