Added horizontal lines
This time, using horizontal lines can divide up a webpage into sections so that it is easier to differentiate the content. This means removing redundant line breaks (or <br>) in some parts of the webpages.
The paragraph text in the course pages has been moved over to the right by shifting the margin over by 25 pixels for better reading.
The <nav> tag has also been added in the About page. More to come for this part...