Make some courses inaccessible
Inside the CSS file, I set the pointer-events property for inaccessible links to none to indicate that the users can't access the link yet. In addition, I also replaced the "under construction" pages for the courses not yet started with the actual pages. Said pages are not yet accessible directly, but don't worry. These will be accessible in the future.