Split up the timeline
At the current state, while it is easy to see the list of courses taken by semester and year, it becomes difficult to follow along as the number of semesters increases. This holds true when trying to figure out which semesters are taken in bachelor's and in master's.
Alleviate this problem by dividing the buttons into smaller parts: the larger one for bachelor's and the smaller one for master's.
(One more thing: set the buttons together as a flex instead of an inline-table, and add the text to better differentiate the group of buttons.)