Some minor adjustments
In the JavaScript file that displays last time and date visited, append a leading 0 before minute, but only if the current minute is less than 10.
In the React Basics page, split up the text into bullet points for better readability.
Images for the React calculator app has been added. The image gallery for React Basics will be added in the next commit.