Major changes for image gallery
The left and right buttons have been removed from implementation, due to the positioning problems with the image gallery. The dots that navigates over to the corresponding images should suffice, so there's no need for extra redundancy in the slideshow content.
Also, the "Under Construction" page has been removed along with the stylesheet. There's no need to keep that as the pages are already implemented though not yet accessible.