Major changes for the webpage
Apply the styling for both the webpages of the Bachelor's and Master's courses list based on the ones applied for the Certificates page. With all of the courses grouped based on the corresponding semester, now it should be easier to follow along.
Also, trim out the excessive background space for the expandable content by setting the max-width to the appropriate values.
For the certificates page, simplify the status text by clipping out the wordiness.